Update thread.c
s

+++ src/threads/thread.c
@@ -98,7 +98,7 @@ init_thread (struct thread *t, const char *name, int priority) {
     memset (t, 0, sizeof *t);
     t->status = THREAD_BLOCKED;
     strlcpy (t->name, name, sizeof t->name);
-    t->tf.esp = (uint64_t) t + PGSIZE - sizeof (void *);
+    t->tf.esp = (uint32_t)t + PGSIZE - sizeof(void *);
     t->priority = priority;
     t->magic = THREAD_MAGIC;
     // priority donation
